Extract SystemTheme class
